requirements.txt
# packages in environment at C:\Users\sun\anaconda3\envs\test:
#
# Name                    Version                   Build  Channel
# alphapose                 0.3.0+038435e             dev_0    <develop>
blas                      1.0                         mkl  
certifi                   2020.6.20                py36_0  
cffi                      1.14.2           py36h7a1dbc1_0  
chardet                   3.0.4                    pypi_0    pypi
cudatoolkit               10.0.130                      0  
cycler                    0.10.0                   pypi_0    pypi
cython                    0.29.21                  pypi_0    pypi
cython-bbox               0.1.3                    pypi_0    pypi
easydict                  1.9                      pypi_0    pypi
freetype                  2.10.2               hd328e21_0  
idna                      2.10                     pypi_0    pypi
intel-openmp              2020.2                      254  
jpeg                      9b                   hb83a4c4_2  
jsonpatch                 1.26                     pypi_0    pypi
jsonpointer               2.0                      pypi_0    pypi
kiwisolver                1.2.0                    pypi_0    pypi
libpng                    1.6.37               h2a8f88b_0  
libtiff                   4.1.0                h56a325e_1  
lz4-c                     1.9.2                h62dcd97_1  
matplotlib                3.3.1                    pypi_0    pypi
mkl                       2020.2                      256  
mkl-service               2.3.0            py36hb782905_0  
mkl_fft                   1.1.0            py36h45dec08_0  
mkl_random                1.1.1            py36h47e9c7a_0  
munkres                   1.1.2                    pypi_0    pypi
natsort                   7.0.1                    pypi_0    pypi
ninja                     1.10.1           py36h7ef1ec2_0  
numpy                     1.19.1           py36h5510c5b_0  
numpy-base                1.19.1           py36ha3acd2a_0  
olefile                   0.46                     py36_0  
opencv-python             4.4.0.42                 pypi_0    pypi
pillow                    7.2.0            py36hcc1f983_0  
pip                       20.2.2                   py36_0  
protobuf                  4.0.0rc2                 pypi_0    pypi
pycocotools               2.0                      pypi_0    pypi
pycparser                 2.20                       py_2  
pyparsing                 3.0.0a2                  pypi_0    pypi
python                    3.6.10               h9f7ef89_2  
python-dateutil           2.8.1                    pypi_0    pypi
pytorch                   1.1.0           py3.6_cuda100_cudnn7_1    pytorch
pyyaml                    5.3.1                    pypi_0    pypi
pyzmq                     19.0.2                   pypi_0    pypi
requests                  2.24.0                   pypi_0    pypi
scipy                     1.1.0                    pypi_0    pypi
setuptools                49.6.0                   py36_0  
six                       1.15.0                     py_0  
sqlite                    3.33.0               h2a8f88b_0  
tensorboardx              2.1                      pypi_0    pypi
timm                      0.1.20                   pypi_0    pypi
tk                        8.6.10               he774522_0  
torchvision               0.3.0              py36_cu100_1    pytorch
tornado                   6.0.4                    pypi_0    pypi
tqdm                      4.48.2                   pypi_0    pypi
urllib3                   1.25.10                  pypi_0    pypi
vc                        14.1                 h0510ff6_4  
visdom                    0.1.8.9                  pypi_0    pypi
vs2015_runtime            14.16.27012          hf0eaf9b_3  
websocket-client          0.57.0                   pypi_0    pypi
wheel                     0.35.1                     py_0  
wincertstore              0.2              py36h7fe50ca_0  
xz                        5.2.5                h62dcd97_0  
zlib                      1.2.11               h62dcd97_4  
zstd                      1.4.5                h04227a9_0  
